+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Change Log
 
+## Version 3.0.2
+
+### 2019-07-24
+
+1. Add copy back
+2. Merge daovoice into `usr.js`
+3. Add copy btn
+4. Add markdown switch
+5. Update success page's clipboard include
+6. Version dump to `3.0.2`
+7. Remove `prismjs`
+
+### 2019-07-23
+
+1. Remove prismjs
+2. Add highlightjs
+3. Add line numbers to highlightjs
+4. Bug fix of duplicate syntax when language change
+5. Update baseUrl, remove `protocol` & `base_url` from config.json
+6. Update highlight's register
+
+### 2019-07-22
+
+1. Add API prefetch
+
+### 2019-07-18
+
+1. Remove syntax on plain
+2. Make font size bigger when view the code & text
+3. No copy when plain yetg
+
 ## Version 3.0.1
 
 ### 2019-07-10